+++ apps/openmw/CMakeLists.txt
@@ -29,7 +29,7 @@ add_openmw_dir (mwinput
     )
 
 add_openmw_dir (mwlua
-    luamanager luascript dynamicluaobject luatimer luautil tes3utillua baseevent genericevent disableableevent disableableeventmanager loadedgame
+    luamanager script dynamicluaobject timer util tes3util baseevent genericevent disableableevent disableableeventmanager loadedgame
     )
 
 add_openmw_dir (mwlua/structs
