+++ apps/openmw/mwlua/functions/sound.cpp
@@ -3,7 +3,7 @@
 #include <extern/sol2/sol.hpp>
 
 #include "../luamanager.hpp"
-#include "../luautil.hpp"
+#include "../util.hpp"
 
 #include "../../mwbase/environment.hpp"
 #include "../../mwbase/soundmanager.hpp"
