+++ apps/openmw/mwlua/tes3util.cpp
@@ -1,4 +1,4 @@
-#include "tes3utillua.hpp"
+#include "tes3util.hpp"
 
 #include "functions/ai.hpp"
 #include "functions/gui.hpp"
