+++ apps/openmw/mwlua/tes3util.hpp
