+++ apps/openmw/mwlua/timer.hpp
