+++ apps/openmw/mwlua/util.cpp
@@ -1,4 +1,4 @@
-#include "luautil.hpp"
+#include "util.hpp"
 
 #include <components/debug/debuglog.hpp>
