+++ apps/openmw/mwlua/util.hpp
